# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Poultry Farm Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Farm {
  id: number;
  name: string;
  nameAr: string;
  houseCount: number;
  activeCycles?: number;
  totalChicks?: number;
}

export interface House {
  id: number;
  farmId: number;
  name: string;
  nameAr: string;
  areaM2: number;
  cycleCount?: number;
}

export type CycleBreed = (typeof CycleBreed)[keyof typeof CycleBreed];

export const CycleBreed = {
  Ross: "Ross",
  Cobb: "Cobb",
  Other: "Other",
} as const;

export type CycleStatus = (typeof CycleStatus)[keyof typeof CycleStatus];

export const CycleStatus = {
  active: "active",
  completed: "completed",
} as const;

export interface Cycle {
  id: number;
  houseId: number;
  cycleNumber: number;
  housingDate: string;
  chickCount: number;
  chickPricePerUnit: number;
  breed: CycleBreed;
  totalFeedKg?: number;
  feedCostTotal?: number;
  totalMedicationCost?: number;
  totalMortality?: number;
  finalWeightKg?: number;
  saleAgedays?: number;
  salePricePerKg?: number;
  otherCosts?: number;
  status: CycleStatus;
  mortalityRate?: number;
  fcr?: number;
  averageWeightKg?: number;
  totalCost?: number;
  totalRevenue?: number;
  netProfit?: number;
  costPerLiveKg?: number;
  profitPerChick?: number;
}

export type CreateCycleBodyBreed =
  (typeof CreateCycleBodyBreed)[keyof typeof CreateCycleBodyBreed];

export const CreateCycleBodyBreed = {
  Ross: "Ross",
  Cobb: "Cobb",
  Other: "Other",
} as const;

export type CreateCycleBodyStatus =
  (typeof CreateCycleBodyStatus)[keyof typeof CreateCycleBodyStatus];

export const CreateCycleBodyStatus = {
  active: "active",
  completed: "completed",
} as const;

export interface CreateCycleBody {
  cycleNumber: number;
  housingDate: string;
  chickCount: number;
  chickPricePerUnit: number;
  breed: CreateCycleBodyBreed;
  totalFeedKg?: number;
  feedCostTotal?: number;
  totalMedicationCost?: number;
  totalMortality?: number;
  finalWeightKg?: number;
  saleAgedays?: number;
  salePricePerKg?: number;
  otherCosts?: number;
  status: CreateCycleBodyStatus;
}

export interface CompanySummary {
  totalFarms: number;
  totalHouses: number;
  activeCycles: number;
  completedCycles: number;
  totalChicksAllTime: number;
  totalRevenueAllTime: number;
  totalProfitAllTime: number;
  averageFcr: number;
  averageMortalityRate: number;
  bestFarmName?: string;
  worstFarmName?: string;
}

export interface FarmPerformance {
  farmId: number;
  farmName: string;
  farmNameAr: string;
  totalCycles: number;
  totalChicks: number;
  totalRevenue: number;
  totalProfit: number;
  averageFcr: number;
  averageMortalityRate: number;
  profitMarginPercent: number;
  rank: number;
}

export interface HouseStats {
  houseId: number;
  houseName: string;
  houseNameAr: string;
  cycleCount: number;
  totalProfit: number;
  averageFcr: number;
  averageMortalityRate: number;
}

export interface FarmDashboard {
  farm: Farm;
  houses: House[];
  cycles: Cycle[];
  bestCycle?: Cycle;
  worstCycle?: Cycle;
  totalProfit: number;
  averageFcr: number;
  averageMortalityRate: number;
  houseComparison: HouseStats[];
}

export interface MortalityLog {
  id: number;
  cycleId: number;
  logDate: string;
  count: number;
  notes?: string;
  createdAt: string;
}

export interface AddMortalityLogBody {
  logDate: string;
  /** @minimum 1 */
  count: number;
  notes?: string;
}

export interface CycleWithContext {
  cycle: Cycle;
  farmName: string;
  farmNameAr: string;
  houseName: string;
  houseNameAr: string;
}