# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Poultry Farm Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cycles/{cycleId}/mortality:
    get:
      operationId: listMortalityLogs
      tags: [cycles]
      summary: List daily mortality logs for a cycle
      parameters:
        - name: cycleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of mortality logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MortalityLog"
    post:
      operationId: addMortalityLog
      tags: [cycles]
      summary: Add a daily mortality log entry
      parameters:
        - name: cycleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMortalityLogBody"
      responses:
        "201":
          description: Log created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MortalityLog"
  /cycles/{cycleId}/mortality/{logId}:
    delete:
      operationId: deleteMortalityLog
      tags: [cycles]
      summary: Delete a mortality log entry
      parameters:
        - name: cycleId
          in: path
          required: true
          schema:
            type: integer
        - name: logId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /farms/{farmId}/dashboard:
    get:
      operationId: getFarmDashboard
      tags: [dashboard]
      summary: Get dashboard data for a specific farm
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Farm dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmDashboard"
...
components:
  schemas:
...
    Farm:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        houseCount:
          type: integer
        activeCycles:
          type: integer
        totalChicks:
          type: integer
      required: [id, name, nameAr, houseCount]
    House:
      type: object
      properties:
        id:
          type: integer
        farmId:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        areaM2:
          type: number
        cycleCount:
          type: integer
      required: [id, farmId, name, nameAr, areaM2]
    Cycle:
      type: object
      properties:
        id:
          type: integer
        houseId:
          type: integer
        cycleNumber:
          type: integer
        housingDate:
          type: string
          format: date
        chickCount:
          type: integer
        chickPricePerUnit:
          type: number
        breed:
          type: string
          enum: [Ross, Cobb, Other]
        totalFeedKg:
          type: number
        feedCostTotal:
          type: number
        totalMedicationCost:
          type: number
        totalMortality:
          type: number
        finalWeightKg:
          type: number
        saleAgedays:
          type: integer
        salePricePerKg:
          type: number
        otherCosts:
          type: number
        status:
          type: string
          enum: [active, completed]
        mortalityRate:
          type: number
        fcr:
          type: number
        averageWeightKg:
          type: number
        totalCost:
          type: number
        totalRevenue:
          type: number
        netProfit:
          type: number
        costPerLiveKg:
          type: number
        profitPerChick:
          type: number
      required: [id, houseId, cycleNumber, housingDate, chickCount, chickPricePerUnit, breed, status]
...
    FarmDashboard:
      type: object
      properties:
        farm:
          $ref: "#/components/schemas/Farm"
        houses:
          type: array
          items:
            $ref: "#/components/schemas/House"
        cycles:
          type: array
          items:
            $ref: "#/components/schemas/Cycle"
        bestCycle:
          $ref: "#/components/schemas/Cycle"
        worstCycle:
          $ref: "#/components/schemas/Cycle"
        totalProfit:
          type: number
        averageFcr:
          type: number
        averageMortalityRate:
          type: number
        houseComparison:
          type: array
          items:
            $ref: "#/components/schemas/HouseStats"
      required: [farm, houses, cycles, totalProfit, averageFcr, averageMortalityRate, houseComparison]
    HouseStats:
      type: object
      properties:
        houseId:
          type: integer
        houseName:
          type: string
        houseNameAr:
          type: string
        cycleCount:
          type: integer
        totalProfit:
          type: number
        averageFcr:
          type: number
        averageMortalityRate:
          type: number
      required: [houseId, houseName, houseNameAr, cycleCount, totalProfit, averageFcr, averageMortalityRate]
    MortalityLog:
      type: object
      properties:
        id:
          type: integer
        cycleId:
          type: integer
        logDate:
          type: string
          format: date
        count:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
      required: [id, cycleId, logDate, count, createdAt]
    AddMortalityLogBody:
      type: object
      properties:
        logDate:
          type: string
          format: date
        count:
          type: integer
          minimum: 1
        notes:
          type: string
      required: [logDate, count]
...